Reconfigure after joining Ruby/Scala repositories
Before this change, there were two repositories: one that was an attempt
at working in Ruby, and another that was Scala focused. The Scala
repository did not have any database information: it relied on the Ruby
repository. After working for a while with Scala, I chose to merge the
two projects as one. The project had to be slightly reconfigured in
order to make the build work from within the Vagrant VM.