Record which transaction implements a bank account transaction
When we import a batch of bank account transactions, the values aren't
yet stored in the "real" accounting tables: they're in a staging area,
where they'll be cleaned up and imported. Record which transaction entry
implements which bank account transaction, in order to indicate to the
user which bank account transactions still need to be processed.