Improve the developer experience by automatically recompiling
Why not use the Elm reactor? Because I need to expose two different servers on the
same HTTP port. I could run nginx in front of both and make them both appear as the
same physical server, but I'm lazy. For production use, this will have to be done
anyway. Nginx will be perfect to serve the few static files we have anyway.