Begin to save accounts to the database
This is ugly, but it "works": the form data is sent to the server as multipart/form-data.
This is way more work than is required. I need to have better control over the data from
Elm, but I'm not there yet. There is skeleton code to implement POSTing JSON data, but
it's not ready yet.