Let routes dictate the transactional behaviour
Some routes are read-only and do not require transactional behaviour. Others need
to write, and thus require full transactional behaviour. Putting that information
in the route seems to make sense, although putting the same information on the
controller itself might be preferable. I'll punt for now and think about it.