Add ability to upload bank account transactions to server
There was much angst related to the FrontController. Ideally, the form's data
could be uploaded using XMLHttpRequest, but this proved to be slightly more
difficult. Instead, I took the pragmatic approach and let the client choose
where to redirect the browser. If using curl or some other method, a plain
POST will do the right thing and return straight JSON.